+++ common/src/main/java/dev/ithundxr/createnumismatics/registry/NumismaticsBlockEntities.java
@@ -21,6 +21,7 @@ package dev.ithundxr.createnumismatics.registry;
 import com.simibubi.create.foundation.data.CreateRegistrate;
 import com.tterrag.registrate.util.entry.BlockEntityEntry;
 import dev.ithundxr.createnumismatics.Numismatics;
+import dev.ithundxr.createnumismatics.base.data.BuilderTransformers;
 import dev.ithundxr.createnumismatics.content.bank.blaze_banker.BlazeBankerBlockEntity;
 import dev.ithundxr.createnumismatics.content.bank.blaze_banker.BlazeBankerRenderer;
 import dev.ithundxr.createnumismatics.content.depositor.AndesiteDepositorBlockEntity;
@@ -43,6 +44,7 @@ public class NumismaticsBlockEntities {
 
     public static final BlockEntityEntry<BrassDepositorBlockEntity> BRASS_DEPOSITOR = REGISTRATE.blockEntity("brass_depositor", BrassDepositorBlockEntity::new)
         .validBlocks(NumismaticsBlocks.BRASS_DEPOSITOR)
+        .transform(BuilderTransformers.discreteCoinBagStorage())
         .register();
 
     public static final BlockEntityEntry<BlazeBankerBlockEntity> BLAZE_BANKER = REGISTRATE.blockEntity("blaze_banker", BlazeBankerBlockEntity::new)