+++ fabric/src/main/java/dev/ithundxr/createnumismatics/fabric/NumismaticsImpl.java
@@ -26,7 +26,6 @@ import dev.ithundxr.createnumismatics.registry.commands.arguments.EnumArgument;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.ArgumentTypeRegistry;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
-import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.resources.ResourceLocation;